Update and rename Sentinel 2 to Sentinel 2 Aotea Composites
Primary cloud masking is does using QA bands and cloud score calculation with a buffer to block the cirrus edges of clouds. Gaps are filled in (using mosaic) with a less stringent, but equally cloud free image that was cloud masked using the derived 'cloud probability' band. 

NOTE: For loop allow multiple years to be exported at once, but I recommend doing on at a time due to heavy compute for GEE.